Bug fixes and Additions
Fixed a bug that was causing a graphical error on firefox and IE.
Fixed a bug where users could not click on the label in a form to activate the input field.
Added comments to the html and css files.
Changed the way the social media icons where styled to make them cross-browser and not hard coded.
Deleted bloat code.